Reorganize plugins somewhat
